backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
langchain
langchain-openai
langchain-chroma
langchain-community
pypdf==3.17.1
python-dotenv==1.0.0
pydantic
chromadb
openai
